requirements.txt
PyQt6>=6.6.0
PyMuPDF>=1.23.0
transformers>=4.35.0
torch>=2.1.0
networkx>=3.2
nltk>=3.8
scikit-learn>=1.3.0
bitsandbytes>=0.41.0
python-docx>=1.1.0
reportlab>=4.0.0
pyinstaller>=6.0.0

